third_party/blink/web_tests/html5lib/generated/run-tests7-write-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Unexpected token '<'
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select><select>. Please add the missing </select> end tag.
../resources/tests7.dat:
13
14
30

Test 13 of 30 in ../resources/tests7.dat failed. Input:
<!doctype html><select><input>X
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|       <input>
|       "X"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     <input>
|     "X"

Test 14 of 30 in ../resources/tests7.dat failed. Input:
<!doctype html><select><select>X
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     <select>
|       "X"
Expected:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <select>
|     "X"

Test 30 of 30 in ../resources/tests7.dat failed. Input:
<select><keygen>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|       <keygen>
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
